requirements.txt
langchain-community>=0.0.10
langchain-openai>=0.0.5
python-dotenv>=1.0.0
mysql-connector-python>=8.0.0
rich>=13.0.0
fastapi>=0.68.0
uvicorn>=0.15.0
jinja2>=3.0.1
python-multipart>=0.0.5
